Create find duplicates in an array
Find duplicates in an array

Given an array a of size N which contains elements from 0 to N-1, you need to find all the elements occurring more than once in the given array. Return the answer in ascending order. If no such element is found, return list containing [-1].

Example 1:
Input:
N = 4
a[] = {0,3,1,2}
Output:
-1
Explanation: There is no repeating element in the array. Therefore output is -1.